+++ metadata.rb
@@ -1,8 +1,8 @@
-name             'mokhan-myface'
+name             'stronglifters-rails'
 maintainer       'mo khan'
 maintainer_email 'mo@mokhan.ca'
 license          'All rights reserved'
-description      'Installs/Configures mokhan-myface'
+description      'Installs/Configures stronglifters-rails'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.0'