+++ recipes/default.rb
@@ -1,8 +1,8 @@
 include_recipe "apt"
 include_recipe "build-essential"
-include_recipe "mokhan-myface::user"
-include_recipe "mokhan-myface::aws"
-include_recipe "mokhan-myface::monit"
+include_recipe "stronglifters-rails::user"
+include_recipe "stronglifters-rails::aws"
+include_recipe "stronglifters-rails::monit"
 
 packages = [
   "curl",